# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Project workspace API for the Pitch Intelligence Workspace
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /inventory/sync:
    post:
      operationId: runInventorySync
      tags: [inventory]
      summary: Synchronize Single and Double shelter tabs from the configured live Google Sheet
      responses:
        "200":
          description: Completed sync run
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InventorySyncRun" }
        "409": { description: Sync already running or source not configured }
        "502": { description: Google Sheets read failed; no partial inventory changes committed }
...
components:
  schemas:
...
    InventorySyncChange:
      type: object
      properties:
        id: { type: string, format: uuid }
        syncRunId: { type: string, format: uuid }
        inventoryAssetId: { type: ["string", "null"], format: uuid }
        sourceKey: { type: string }
        changeType: { type: string }
        changedFields: { type: object, additionalProperties: true }
        createdAt: { type: string, format: date-time }
      required: [id, syncRunId, inventoryAssetId, sourceKey, changeType, changedFields, createdAt]
    InventorySyncRun:
      type: object
      properties:
        id: { type: string, format: uuid }
        sourceFamily: { type: string }
        status: { type: string }
        trigger: { type: string }
        sourceTitle: { type: ["string", "null"] }
        sourceUrl: { type: ["string", "null"] }
        sourceRows: { type: integer }
        addedCount: { type: integer }
        updatedCount: { type: integer }
        unchangedCount: { type: integer }
        removedCount: { type: integer }
        missingCount: { type: integer }
        reviewCount: { type: integer }
        duplicateCount: { type: integer }
        error: { type: ["string", "null"] }
        startedAt: { type: string, format: date-time }
        completedAt: { type: ["string", "null"], format: date-time }
        changes: { type: array, items: { $ref: "#/components/schemas/InventorySyncChange" } }
      required: [id, sourceFamily, status, trigger, sourceTitle, sourceUrl, sourceRows, addedCount, updatedCount, unchangedCount, removedCount, missingCount, reviewCount, duplicateCount, error, startedAt, completedAt]
...